+++ src/test/groovy/com/blackbuild/groovy/configdsl/transform/InheritanceSpec.groovy
@@ -144,6 +144,91 @@ class InheritanceSpec extends AbstractDSLSpec {
         thrown(MultipleCompilationErrorsException)
     }
 
+    @Issue("130")
+    def "legal: abstract parent class defines no key, but child defines key"() {
+        when:
+        createClass('''
+            package pk
+
+            @DSL
+            abstract class Foo {
+                String name
+                String parentValue
+            }
+
+            @DSL
+            class Bar extends Foo {
+                @Key String value
+            }
+        ''')
+
+        then:
+        notThrown(MultipleCompilationErrorsException)
+
+        when:
+        instance = getClass("pk.Bar").create("bla") {
+            name "Kurt"
+        }
+
+        then:
+        noExceptionThrown()
+    }
+
+    @Issue("130")
+    def "abstract non-keyed superclass can access key of subclass"() {
+        given:
+        createClass('''
+            package pk
+
+            @DSL
+            abstract class Foo {
+                abstract String getId()
+
+                String getUniqueId() {
+                    return "$id-abc"
+                }             
+            }
+
+            @DSL
+            class Bar extends Foo {
+                @Key String id
+            }
+        ''')
+
+        when:
+        instance = getClass("pk.Bar").create("bla")
+
+        then:
+        instance.uniqueId == "bla-abc"
+
+
+    }
+
+    @Issue("130")
+    def "error: abstract non-abstract abstract hierarchy"() {
+        when:
+        createClass('''
+            package pk
+
+            @DSL
+            abstract class Foo {
+                String name
+                String parentValue
+            }
+
+            @DSL
+            class Bar extends Foo {
+            }
+            @DSL
+            abstract class Boo extends Bar {
+                @Key String value
+            }
+        ''')
+
+        then:
+        thrown(MultipleCompilationErrorsException)
+    }
+
     def "error: parent class defines key, child defines different key"() {
         when:
         createClass('''